Separate `ContextDocstrings` from `Context` and make it pluggable
