+++ dottydoc/src/dotty/tools/dottydoc/util/syntax.scala
@@ -2,20 +2,15 @@ package dotty.tools
 package dottydoc
 package util
 
-import dotc.core.Contexts.{ Context, DocBase }
+import dotc.core.Contexts.Context
+import dotc.core.Comments._
 import model.Package
+import core.ContextDottydoc
 
 object syntax {
-  implicit class RichDocContext(val ctx: Context) extends AnyVal {
-    def docbase: DocBase = ctx.getDocbase getOrElse {
+  implicit class ContextWithContextDottydoc(val ctx: Context) extends AnyVal {
+    def docbase: ContextDottydoc = ctx.docCtx.getOrElse {
       throw new IllegalStateException("DocBase must be set before running dottydoc phases")
-    }
-  }
-
-  implicit class RichDocBase(val db: DocBase) {
-    def packages: Map[String, Package] = db.packagesAs[Package].toMap
-
-    def packagesMutable: collection.mutable.Map[String, Package] =
-      db.packagesAs[Package]
+    }.asInstanceOf[ContextDottydoc]
   }
 }
